+++ NEWS.md
@@ -7,6 +7,7 @@
 - Added range dseq for jobs over ranges of integers.
 - Added dval dseq for jobs over the content of a dval.
 - Correct local processing of zero-split dseqs.
+- Ensure output paths exists on job success.
 
 ## 0.6.1 / 2014-10-21